requirements.txt
# AI Training and Inference
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Model Serving
vllm>=0.2.6
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
scikit-learn>=1.3.0

# Web Interface
gradio>=4.0.0
streamlit>=1.28.0

# Utilities
requests>=2.31.0
loguru>=0.7.0
pyyaml>=6.0
tensorboard>=2.14.0

# ModelScope
modelscope>=1.9.0